RAG/RAG_API_Clean_Redis/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
numpy>=1.24
pandas>=2.0
qdrant-client>=1.7
python-dotenv>=1.0
ollama>=0.2
openai>=1.40.0
google-genai>=0.3.0

redis>=5.0.0
